requirements.txt
# ----------------------
# Flask
# ----------------------
flask

# ----------------------
# Requests
# ----------------------
requests

# ----------------------
# Wget
# ----------------------
wget

tensorflow

Pillow
